refactor: Facade 패턴을 이용해 내부 로직을 드러내지 않고 메서드만 사용하도록 변경
